+++ client/components/SingleItemView.js
@@ -6,7 +6,7 @@ import MapSingleItem from './MapSingleItem'
 import MobileFooter from './MobileFooter'
 
 // Render functional
-const singleView = (props) => {
+const SingleItemView = (props) => {
   console.log(props)
   return (
     <div className="container-fluid mb-4 p-0" style={{height: '90vh'}}>
@@ -178,4 +178,4 @@ const singleView = (props) => {
 //   },
 // })
 
-export default singleView
+export default SingleItemView
