+++ client/components/index.js
@@ -12,7 +12,7 @@ export {Login, Signup} from './AuthForm'
 export {default as Items} from './Items'
 export {default as temp} from './temp'
 export {default as NewItemForm} from './NewItemForm'
-export {default as SingleView} from './singleView'
+export {default as SingleItemView} from './SingleItemView'
 export {default as AllUsers} from './AllUsers'
 export {default as UserView} from './UserView'
 export {default as User} from './User'
