+++ client/routes.js
@@ -12,7 +12,7 @@ import {
   temp,
   NewItemForm,
   Items,
-  SingleView,
+  SingleItemView,
   AllUsers,
   User,
 } from './components'
@@ -36,7 +36,7 @@ class Routes extends Component {
         <Route path="/items" component={Items} />
         <Route path="/mapallitems" component={MapAllItems} />
         <Route path="/mapsingleitem" component={MapSingleItem} />
-        <Route path="/singleview" component={SingleView} />
+        <Route path="/singleview" component={SingleItemView} />
         <Route path="/temp" component={temp} />
         <Route path="/login" component={Login} />
         <Route path="/signup" component={Signup} />
